+++ includes/class.llms.person.handler.php
@@ -1,11 +1,15 @@
 <?php
-defined( 'ABSPATH' ) || exit;
-
 /**
  * User Handling for login and registration (mostly)
  *
  * @since    3.0.0
- * @version  3.24.0
+ * @version  [version]
+ */
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * LLMS_Person_Handler class.
  */
 class LLMS_Person_Handler {
 
@@ -625,61 +629,52 @@ class LLMS_Person_Handler {
 
 	/**
 	 * Login a user
-	 * @param    array     $data    array of login data
-	 * @return   mixed              WP_Error on error or the ID of the user
+	 *
+	 * @param    array $data array of login data.
+	 * @return   WP_Error|int WP_Error on error or the WP_User ID.
 	 * @since    3.0.0
-	 * @version  3.0.0
+	 * @version  [version]
 	 */
 	public static function login( $data ) {
 
 		do_action( 'lifterlms_before_user_login', $data );
 
-		// validate the fields & allow custom validation to occur
+		// validate the fields & allow custom validation to occur.
 		$valid = self::validate_fields( apply_filters( 'lifterlms_user_login_data', $data ), 'login' );
 
-		// if errors found, return them
+		// if errors found, return them.
 		if ( is_wp_error( $valid ) ) {
-
-			return apply_filters( 'lifterlms_user_login_errors', $valid, $data );
-
-		} // End if().
-		else {
-
-			$creds = array();
-			$e = new WP_Error( 'login-error', __( 'Could not find an account with the supplied email address and password combination.', 'lifterlms' ) );
-
-			// get the username from the email address
-			if ( 'yes' === get_option( 'lifterlms_registration_generate_username' ) && apply_filters( 'lifterlms_get_username_from_email', true ) ) {
-
-				$user = get_user_by( 'email', $data['llms_login'] );
-
-				if ( isset( $user->user_login ) ) {
-
-					$creds['user_login'] = $user->user_login;
-
-				} else {
-
-					return $e;
-
-				}
-			} else {
-
-				$creds['user_login'] = $data['llms_login'];
-
-			}
-
-			$creds['user_password'] = $_POST['llms_password'];
-			$creds['remember'] = isset( $_POST['llms_remember'] );
-
-			$signon = wp_signon( apply_filters( 'lifterlms_login_credentials', $creds ), is_ssl() );
-
-			if ( is_wp_error( $signon ) ) {
-				$e = apply_filters( 'lifterlms_user_login_errors', $e, $data );
-				return $e;
-			} else {
-				return $signon->ID;
-			}
+			return apply_filters( 'lifterlms_user_login_errors', $valid, $data, false );
 		}
+
+		$creds = array();
+		$creds['user_login'] = $data['llms_login'];
+
+		$err = new WP_Error( 'login-error', __( 'Could not find an account with the supplied email address and password combination.', 'lifterlms' ) );
+
+		// get the username from the email address
+		if ( llms_parse_bool( get_option( 'lifterlms_registration_generate_username' ) ) && apply_filters( 'lifterlms_get_username_from_email', true ) ) {
+
+			$user = get_user_by( 'email', wp_unslash( $data['llms_login'] ) );
+
+			if ( ! isset( $user->user_login ) ) {
+				return apply_filters( 'lifterlms_user_login_errors', $err, $data, false );
+			}
+
+			$creds['user_login'] = $user->user_login;
+
+		}
+
+		$creds['user_password'] = $data['llms_password'];
+		$creds['remember'] = isset( $data['llms_remember'] );
+
+		$signon = wp_signon( apply_filters( 'lifterlms_login_credentials', $creds ), is_ssl() );
+
+		if ( is_wp_error( $signon ) ) {
+			return apply_filters( 'lifterlms_user_login_errors', $err, $data, $signon );
+		}
+
+		return $signon->ID;
 
 	}